Reimplement portmapper authentication.
Portmapper authentication also saves the connecting client pid's start
time, then adds an additional step: portmapper sends a socket back to the
client, that replaces the initial connection, and the client must resend its
credentials back over the new socket. Portmapper verifies the client's pid,
executable, and process start time, hasn't changed. This eliminates potentially
exploitable race conditions.

The additional step of completing a second round of credentials checks over
a socket that the client could only possible have after portmapper verified
its initial credentials, together with the requirements that the process's
start time hasn't changed, pretty much prevents the client from playing
games with forking, keeping the connection in the child process, and the
original process re-execing the executable the process wants to impersonate,
hoping to do that before portmapper grabs the pid's executable.

Reimplement lookup of the pid's executable on freebsd. /proc/curproc/file
cannot be relied from, it is populated from a kernel cache that's subject to
purging. The workaround is to use sysctl to grab the executable's dev and
inode, which is the only thing that can be relied on. Then, on freebsd, also
grab the command args, also using sysctl, then search for argv[0] in the PATH,
stat-ing it, and verifying that it matches the dev+ino from sysctl. On Linux,
read /proc/self/exe as usual.

On freebsd, the process sends its derived executable pathname on the initial
portmapper connection, and the portmapper verifies that stat-ing it gives
the process's dev/ino that's still available from sysctl. This fixes the
caching issue.

This removes all dependencies of /proc on freebsd, replacing it with sysctl.
The configure tests are simplified accordingly.

Also, fix the addresses in backtraces on freebsd.

This removes most of the contents of pidinfo.H, since the executable path
lookup diverged significantly on freebsd. What was there gets moved into
the portmapper daemon, which continues to work, as it was, on Linux.