Client-side HTTP authorization cache.
Initial implementation that uses basic authorization scheme, but define
an abstract clientauthObj superclass for authorization scheme drivers, with
clientauth::base::create_basic() creating a basic scheme driver.

clientauth->add_header() invokes the driver to add authorization headers to
an HTTP request.

clientauthcacheObj implements the cache for proxy and www authorizations.

save_authorization() takes the original resp with authentication
challenges, and the original request URI, a constructed clientauth, and puts
it in the cache.

search_authorization() searches the cache for known authorizations for the
given request, and invokes add_header() for each found authorization.