Add an authorization cache to http useragent object.
Store the request URI in useragent::base::response.

Add discardbody() to useragent::base::response, that explicitly discards the
body of the response.

do_request() renamed to do_request_with_auth(). The replacement do_request()
handles authentication:

- If the request URI has an explicit authority userinfo, dummy up a fake
authentication-required response, and use it to cache a basic authentication
client for an empty realm. Remove the userinfo from the URI's authority.

- Search the authorization cache for any authorizations whose protection
space encompasses the request's URI, and add their headers to the request.

- Call do_request_with_auth(), then check if the response contained any
authentication challenges. If it had, check if those challenges can be
found in the authentication cache. If all of them were, discard the
first request's body, and try again.

- Store challenges that were not filled from the cache, challenges that
the caller must process, in convenient, parsed form, in the returned
response object.

Add set_authorization(): once the client responds to a challenge, save the
new authorization in the cache.

Implement all of this for basic authorization.