A template for a hierarchical key container.
Imlpements a thread-safe container that uses a hierarchical key.

A typical key would be, say std::list<std::string>.

You can search for a key, and if it does not exist, you'll find some other
key that matches the prefix of your key, if it exists. You'll find the closest
parent to the nonexisting key.