Simplify how authentication implementations are searched for.
get_default_protection_space() returns two search paths, a realm-based search
path and a URI-based search path.

The uri search for basic auth modules is "uri" for proxy auth and
[authority]/"uri" for www auth.

The uri search for digest auth modules is "uri"/[authority]/[path] for proxy
auth and [authority]/"uri"/ for www auth.

Dump code from search_authorizations(), simplifying it. Split it into
search_authorization_path().

fail_authorization() does not need the authorizations parameter.

fail_authorization() simply uses the realm path to find the authentication
module. Then, the existing code calls get_protection space, and finds all the
places this module gets installed into.

This gets rid of the special case logic in basic's get_protection_space(),
it gets essentially incorporated into get_default_protection_space().
basic::get_protection_space simply returns both the realm and the uri
search path.

Remove the code that fails the default empty realm from fail_authorization(),
and move it to the user agent object. The user agent is the one that installs
the userinfo portion of the URL as the empty realm, and it can deal with it
itself, it does not belong in the cache object.