Have headeriter take care of stripping leading newlines.
Remove leading space stripping from headercollector.

After seeing :, do not emit header_contents_start right away, wait for a first
non-space. Implement this by transaction to a new state, in_starting_contents,
that always transition to in_contents, instead of transition to in_contents
as soon as : is seen. As long as a space is seen, remain in_starting_contents.

If a newline is seen, transition to in_starting_contents_newline, then back
to in_starting_contents, where we came from, and resume waiting for the next
non-whitespace char. This makes it possible for in_starting_contents to
see a body_start or eof, so have it fall through to emitting the delayed
in_contents_start, then call start_header() to do that processing, to avoid
duplicating code.