Clean up message catalogs.
Change the constructor parameters. First parameter is the text domain,
all remaining parameters are optional, and the locale defaults to the
global locale.

Remove several overloads, and use string_view to handle them.

Make gettext_helper a lightweight temporary object instead of a heavyweight
reference-counted object. The requires removal of format() and formatn()
from messages, so that gettext_helper only needs to capture a string_view
for the formatting string.