Clean up weak objects.
Get rid of destroyCallbackObj, and simply borrow functionref for this
purpose. The destroy callback is simply stored in a functionref.

Rename weakinfo to weakinfoObj, for consistency, and hide it. weakinfoObj
inherits from weakinfobaseObj which declares get_strongref(). This removes
the need for weakptr to pull in weakinfo and all of its dependencies, all
it needs is get_strongref.

For extra safety, add destroy_aborted flag, and use it, and not just a
condition variable signal, since wait() is documented to possibly have a
spurious wake-up, so the weak object race condition check uses a bog-standard
condition variable+flag.

Move all stuff in weakinfoObj into a mutex-protected object, to prove by
contract that its only accessed under a mutex lock.